app/requirements.txt
requests
flask
python-dotenv
tabulate